[Refactor pt1] - Restructure Restaurant List to separate GUI and listeners away from objects and refactor AddNewRestaurantPanel class 	- Replace new AddNewScreen() with AddNewRestaurantPanel class 	- New Directory Structure in src/RestaurantList/ 		- GUI/ contains GUI code 		- listeners/ contains listeners that are shared between different GUI components 		- verifies/ contains verifies that are shared between different GUI components
